+++ CHANGELOG.md
@@ -32,3 +32,7 @@ December 5th, 2023
 - Removed the interface directory configuration option in favor of automatically detecting it from the connected Predator configuration file.
 - Cortex now kills ALPR processes when stopping Predator.
 - The instance recovery tools now support both Predator Fabric and vanilla Predator.
+- Added advanced management tools.
+    - Predator's real-time ALPR functionality can now be registered as a SystemD service from the Cortex interface.
+    - Cortex now allows users to view the contents of files in the interface, instance, and working directories.
+